requirements.txt
anticaptchaofficial==1.0.43
selenium==4.2.0
pathlib==1.0.1
telebot==0.0.4
requests==2.28.0
PyTelegramBotAPI==4.6.0
environ==1.0


python-dotenv==0.20.0